Rewrite classes/constructor Description section
This is a work in progress. I ran out of steam in the middle of writing it. I think the next step will be to push the concept of default constructors lower and start with explicit constructors first. I think this will make it easier to describe why you need an explicit `super()` call and explicit constructors from derived classes.